+++ damavis-spark-core/src/main/scala/com/damavis/spark/pipeline/stages/CacheStage.scala
@@ -1,4 +1,4 @@
-package com.damavis.spark.pipeline.predefined
+package com.damavis.spark.pipeline.stages
 
 import com.damavis.spark.pipeline.PipelineStage
 import org.apache.spark.sql.DataFrame
